## Overview

This PR uses delete range API recently introduced to `rocksdb::WriteBatch` in order to simplify / improve performance of index removals.

The PR includes a benchmark which locally shows >100x (yep, hundred times) improvement for some test cases:

```text
Clearing a list with 100 x 512-byte values:
    old:   [450.63 us 491.15 us 530.34 us]
    new:   [9.9241 us 10.079 us 10.242 us] (~50x times better!)

Clearing a list with 500 x 512-byte values:
    old:   [2.8341 ms 3.0068 ms 3.1483 ms]
    new:   [12.212 us 12.694 us 13.259 us] (~240x times better!)
```

I don't know whether these benchmarks can be taken at face value (the entire DB is likely to be cached in RAM, etc.), but it likely to have *some* boost in real applications. Another argument for it is that the RAM consumption by `WriteBatch` itself is also reduced from `O(N)` (`N` is the number of keys in the index) to `O(1)`.

---

See: https://jira.bf.local/browse/ECR-3899